+++ wtx-ui/Cargo.toml
@@ -1,7 +1,7 @@
 [dependencies]
 clap = { default-features = false, features = ["derive", "help", "std", "usage"], optional = true, version = "4.0" }
 tokio = { default-features = false, features = ["io-std", "io-util", "macros", "net", "rt-multi-thread"], version = "1.0" }
-wtx = { default-features = false, features = ["tokio"], path = "../wtx", version = "0.41.9" }
+wtx = { default-features = false, features = ["tokio"], path = "../wtx", version = "0.41.10" }
 
 [features]
 default = []
