+++ wtx/Cargo.toml
@@ -121,7 +121,7 @@ name = "wtx"
 readme = "README.md"
 repository = "https://github.com/c410-f3r/wtx"
 rust-version = "1.89"
-version = "0.41.9"
+version = "0.41.10"
 
 [package.metadata.docs.rs]
 all-features = true
