+++ wtx/src/de.rs
@@ -14,8 +14,8 @@ pub mod format;
 mod from_radix_10;
 mod hex;
 mod num_array;
-mod percent_encoding;
 pub mod protocol;
+mod url_encoding;
 
 pub use csv::Csv;
 pub use de_controller::DEController;
@@ -28,7 +28,7 @@ pub use num_array::{
   I8String, I16String, I32String, I64String, U8String, U16String, U32String, U64String, i8_string,
   i16_string, i32_string, i64_string, u8_string, u16_string, u32_string, u64_string,
 };
-pub use percent_encoding::*;
+pub use url_encoding::*;
 
 /// Identifier used to track the number of issued requests.
 pub type Id = u64;
